+++ components/ui-api-layer/Gopkg.lock
@@ -634,6 +634,6 @@
 [solve-meta]
   analyzer-name = "dep"
   analyzer-version = 1
-  inputs-digest = "111ee52a4179e7d55188e1219438238362e19b27a934477ef6811dcd2667cf6e"
+  inputs-digest = "f595076b316799046ca977e2bd30b7d9111f46499156060e3cbe2034411e70fe"
   solver-name = "gps-cdcl"
   solver-version = 1
