+++ components/ui-api-layer/internal/domain/k8s/export_test.go
@@ -21,6 +21,6 @@ func NewSecretResolver(secretGetter v1.SecretsGetter) *secretResolver {
 	return newSecretResolver(secretGetter)
 }
 
-func NewResourceQuotaService(rqInformer cache.SharedIndexInformer, rsInformer cache.SharedIndexInformer, ssInformer cache.SharedIndexInformer, podClient *v1.CoreV1Client) *resourceQuotaService {
+func NewResourceQuotaService(rqInformer cache.SharedIndexInformer, rsInformer cache.SharedIndexInformer, ssInformer cache.SharedIndexInformer, podClient v1.CoreV1Interface) *resourceQuotaService {
 	return newResourceQuotaService(rqInformer, rsInformer, ssInformer, podClient)
 }
