+++ components/ui-api-layer/internal/domain/k8s/resourcequota_resolver_test.go
@@ -11,12 +11,8 @@ import (
 	testingUtils "github.com/kyma-project/kyma/components/ui-api-layer/internal/testing"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	apps "k8s.io/api/apps/v1"
-	api "k8s.io/api/apps/v1beta2"
 	"k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/client-go/informers"
 	"k8s.io/client-go/kubernetes/fake"
 )
@@ -35,7 +31,7 @@ func TestResourceQuotaResolver_ResourceQuotasQuery(t *testing.T) {
 	}, nil)
 	defer lister.AssertExpectations(t)
 
-	resolver := newResourceQuotaResolver(lister, nil, nil, nil, nil)
+	resolver := newResourceQuotaResolver(lister, nil)
 
 	// WHEN
 	result, err := resolver.ResourceQuotasQuery(context.Background(), env)
@@ -44,134 +40,6 @@ func TestResourceQuotaResolver_ResourceQuotasQuery(t *testing.T) {
 	require.NoError(t, err)
 	assert.Len(t, result, 1)
 	assert.Equal(t, gqlschema.ResourceQuota{Name: "mem-default"}, result[0])
-}
-
-func TestResourceQuotaResolver_ResourceQuotaStatus_HardEqualsUsed(t *testing.T) {
-	// GIVEN
-	rqLister := automock.NewResourceQuotaLister()
-	rqLister.On("ListResourceQuotas", fixNamespaceName()).Return(fixResourceQuotaExceeded(), nil)
-	defer rqLister.AssertExpectations(t)
-
-	resolver := newResourceQuotaResolver(rqLister, nil, nil, nil, nil)
-
-	// WHEN
-	status, err := resolver.ResourceQuotaStatus(context.Background(), fixNamespaceName())
-	require.NoError(t, err)
-
-	// THEN
-	assert.True(t, status.Exceeded)
-}
-
-func TestResourceQuotaResolver_ResourceQuotaStatus_ReplicaSetExceeded(t *testing.T) {
-	// GIVEN
-	rqLister := automock.NewResourceQuotaLister()
-	rqLister.On("ListResourceQuotas", fixNamespaceName()).Return(fixResourceQuota(), nil)
-	rsLister := automock.NewReplicaSetLister()
-	rsLister.On("ListReplicaSets", fixNamespaceName()).Return(fixReplicaSetExceeded(), nil)
-	podLister := automock.NewPodsLister()
-	podLister.On("ListPods", fixNamespaceName(), fixReplicaSetMatchLabels()).Return(fixReplicasExceeding(fixReplicaSetMatchLabels()), nil)
-	defer func() {
-		rqLister.AssertExpectations(t)
-		rsLister.AssertExpectations(t)
-		podLister.AssertExpectations(t)
-	}()
-
-	resolver := newResourceQuotaResolver(rqLister, rsLister, nil, podLister, nil)
-
-	// WHEN
-	status, err := resolver.ResourceQuotaStatus(context.Background(), fixNamespaceName())
-	require.NoError(t, err)
-
-	// THEN
-	assert.True(t, status.Exceeded)
-}
-
-func TestResourceQuotaResolver_ResourceQuotaStatus_ReplicaSetExceeded_MaxUnavailable(t *testing.T) {
-	// GIVEN
-	rqLister := automock.NewResourceQuotaLister()
-	rqLister.On("ListResourceQuotas", fixNamespaceName()).Return(fixResourceQuota(), nil)
-	rsLister := automock.NewReplicaSetLister()
-	rsLister.On("ListReplicaSets", fixNamespaceName()).Return(fixReplicaSetWithOwnerReference(), nil)
-	ssLister := automock.NewStatefulSetLister()
-	ssLister.On("ListStatefulSets", fixNamespaceName()).Return([]*apps.StatefulSet{}, nil)
-	defer func() {
-		rqLister.AssertExpectations(t)
-		rsLister.AssertExpectations(t)
-		ssLister.AssertExpectations(t)
-	}()
-
-	client := fake.NewSimpleClientset(fixDeploy())
-	informerFactory := informers.NewSharedInformerFactory(client, 0)
-	informer := informerFactory.Apps().V1beta2().Deployments().Informer()
-	deploySvc := newDeploymentService(informer)
-
-	testingUtils.WaitForInformerStartAtMost(t, time.Second, informer)
-
-	resolver := newResourceQuotaResolver(rqLister, rsLister, ssLister, nil, deploySvc)
-
-	// WHEN
-	status, err := resolver.ResourceQuotaStatus(context.Background(), fixNamespaceName())
-	require.NoError(t, err)
-
-	// THEN
-	assert.False(t, status.Exceeded)
-}
-
-func TestResourceQuotaResolver_ResourceQuotaStatus_ReplicaSetExceeded_MaxUnavailable_Percent(t *testing.T) {
-	// GIVEN
-	rqLister := automock.NewResourceQuotaLister()
-	rqLister.On("ListResourceQuotas", fixNamespaceName()).Return(fixResourceQuota(), nil)
-	rsLister := automock.NewReplicaSetLister()
-	rsLister.On("ListReplicaSets", fixNamespaceName()).Return(fixReplicaSetWithOwnerReference(), nil)
-	podLister := automock.NewPodsLister()
-	podLister.On("ListPods", fixNamespaceName(), fixReplicaSetMatchLabels()).Return(fixReplicasExceeding(fixReplicaSetMatchLabels()), nil)
-	defer func() {
-		rqLister.AssertExpectations(t)
-		rsLister.AssertExpectations(t)
-		podLister.AssertExpectations(t)
-	}()
-
-	client := fake.NewSimpleClientset(fixDeployWithPercentageMaxUnavailable())
-	informerFactory := informers.NewSharedInformerFactory(client, 0)
-	informer := informerFactory.Apps().V1beta2().Deployments().Informer()
-	deploySvc := newDeploymentService(informer)
-
-	testingUtils.WaitForInformerStartAtMost(t, time.Second, informer)
-
-	resolver := newResourceQuotaResolver(rqLister, rsLister, nil, podLister, deploySvc)
-
-	// WHEN
-	status, err := resolver.ResourceQuotaStatus(context.Background(), fixNamespaceName())
-	require.NoError(t, err)
-
-	// THEN
-	assert.True(t, status.Exceeded)
-}
-
-func TestResourceQuotaResolver_ResourceQuotaStatus_StatefulSetExceed(t *testing.T) {
-	// GIVEN
-	rqLister := automock.NewResourceQuotaLister()
-	rqLister.On("ListResourceQuotas", fixNamespaceName()).Return(fixResourceQuota(), nil)
-	rsLister := automock.NewReplicaSetLister()
-	rsLister.On("ListReplicaSets", fixNamespaceName()).Return([]*apps.ReplicaSet{}, nil)
-	ssLister := automock.NewStatefulSetLister()
-	ssLister.On("ListStatefulSets", fixNamespaceName()).Return(fixStatefulSet(), nil)
-	podLister := automock.NewPodsLister()
-	podLister.On("ListPods", fixNamespaceName(), fixStatefulSetMatchLabels()).Return(fixReplicasExceeding(fixReplicaSetMatchLabels()), nil)
-	defer func() {
-		rqLister.AssertExpectations(t)
-		rsLister.AssertExpectations(t)
-		podLister.AssertExpectations(t)
-	}()
-
-	resolver := newResourceQuotaResolver(rqLister, rsLister, ssLister, podLister, nil)
-
-	// WHEN
-	status, err := resolver.ResourceQuotaStatus(context.Background(), fixNamespaceName())
-	require.NoError(t, err)
-
-	// THEN
-	assert.True(t, status.Exceeded)
 }
 
 func TestResourceQuotaResolver_ResourceQuotaStatus_HappyPath(t *testing.T) {
@@ -196,239 +64,15 @@ func TestResourceQuotaResolver_ResourceQuotaStatus_HappyPath(t *testing.T) {
 	informerFactory := informers.NewSharedInformerFactory(client, 0)
 	informer := informerFactory.Apps().V1beta2().Deployments().Informer()
 	deploySvc := newDeploymentService(informer)
+	rqStatusSvc := newResourceQuotaStatusService(rqLister, rsLister, ssLister, podLister, deploySvc)
 
 	testingUtils.WaitForInformerStartAtMost(t, time.Second, informer)
 
-	resolver := newResourceQuotaResolver(rqLister, rsLister, ssLister, podLister, deploySvc)
+	resolver := newResourceQuotaResolver(rqLister, rqStatusSvc)
 
 	// WHEN
 	status, err := resolver.ResourceQuotaStatus(context.Background(), fixNamespaceName())
 	require.NoError(t, err)
 
 	assert.False(t, status.Exceeded)
-}
-
-func fixNamespaceName() string {
-	return "test-ns"
-}
-
-func fixResourceQuotaExceeded() []*v1.ResourceQuota {
-	return []*v1.ResourceQuota{
-		{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "name",
-				Namespace: fixNamespaceName(),
-			},
-			Spec: v1.ResourceQuotaSpec{
-				Hard: v1.ResourceList{
-					v1.ResourceLimitsMemory: resource.MustParse("1Gi"),
-				},
-			},
-			Status: v1.ResourceQuotaStatus{
-				Used: v1.ResourceList{
-					v1.ResourceLimitsMemory: resource.MustParse("1Gi"),
-				},
-			},
-		},
-	}
-}
-
-func fixResourceQuota() []*v1.ResourceQuota {
-	return []*v1.ResourceQuota{
-		{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "name",
-				Namespace: fixNamespaceName(),
-			},
-			Spec: v1.ResourceQuotaSpec{
-				Hard: v1.ResourceList{
-					v1.ResourceLimitsMemory:   resource.MustParse("1Gi"),
-					v1.ResourceRequestsMemory: resource.MustParse("1Gi"),
-				},
-			},
-		},
-	}
-}
-
-func fixStatefulSet() []*apps.StatefulSet {
-	return []*apps.StatefulSet{
-		{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace: fixNamespaceName(),
-			},
-			Spec: apps.StatefulSetSpec{
-				Selector: &metav1.LabelSelector{
-					MatchLabels: fixStatefulSetMatchLabels(),
-				},
-				Replicas: ptrInt32(3),
-			},
-			Status: apps.StatefulSetStatus{
-				Replicas: 2,
-			},
-		},
-	}
-}
-
-func fixReplicaSet() []*apps.ReplicaSet {
-	return []*apps.ReplicaSet{
-		{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace: fixNamespaceName(),
-			},
-			Spec: apps.ReplicaSetSpec{
-				Selector: &metav1.LabelSelector{
-					MatchLabels: fixReplicaSetMatchLabels(),
-				},
-				Replicas: ptrInt32(2),
-			},
-			Status: apps.ReplicaSetStatus{
-				Replicas: 2,
-			},
-		},
-	}
-}
-
-func fixReplicaSetExceeded() []*apps.ReplicaSet {
-	return []*apps.ReplicaSet{
-		{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace: fixNamespaceName(),
-			},
-			Spec: apps.ReplicaSetSpec{
-				Selector: &metav1.LabelSelector{
-					MatchLabels: fixReplicaSetMatchLabels(),
-				},
-				Replicas: ptrInt32(3),
-			},
-			Status: apps.ReplicaSetStatus{
-				Replicas: 2,
-			},
-		},
-	}
-}
-
-func fixReplicaSetWithOwnerReference() []*apps.ReplicaSet {
-	return []*apps.ReplicaSet{
-		{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace: fixNamespaceName(),
-				OwnerReferences: []metav1.OwnerReference{
-					{
-						Name: fixDeployName(),
-						Kind: "Deployment",
-						UID:  "123",
-					},
-				},
-			},
-			Spec: apps.ReplicaSetSpec{
-				Selector: &metav1.LabelSelector{
-					MatchLabels: fixReplicaSetMatchLabels(),
-				},
-				Replicas: ptrInt32(3),
-			},
-			Status: apps.ReplicaSetStatus{
-				Replicas: 2,
-			},
-		},
-	}
-}
-
-func fixReplicaSetMatchLabels() map[string]string {
-	return map[string]string{
-		"label": "replica",
-	}
-}
-
-func fixStatefulSetMatchLabels() map[string]string {
-	return map[string]string{
-		"label": "stateful",
-	}
-}
-
-func fixReplicasExceeding(labels map[string]string) []v1.Pod {
-	return []v1.Pod{
-		{
-			ObjectMeta: metav1.ObjectMeta{
-				Labels: labels,
-			},
-			Spec: v1.PodSpec{
-				Containers: []v1.Container{
-					{
-						Resources: v1.ResourceRequirements{
-							Limits: v1.ResourceList{
-								v1.ResourceMemory: resource.MustParse("2Gi"),
-							},
-							Requests: v1.ResourceList{
-								v1.ResourceMemory: resource.MustParse("2Gi"),
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-}
-
-func fixReplicas(labels map[string]string) []v1.Pod {
-	return []v1.Pod{
-		{
-			ObjectMeta: metav1.ObjectMeta{
-				Labels: labels,
-			},
-			Spec: v1.PodSpec{
-				Containers: []v1.Container{
-					{
-						Resources: v1.ResourceRequirements{
-							Limits: v1.ResourceList{
-								v1.ResourceMemory: resource.MustParse("100Mi"),
-							},
-							Requests: v1.ResourceList{
-								v1.ResourceMemory: resource.MustParse("100Mi"),
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-}
-
-func fixDeploy() *api.Deployment {
-	return &api.Deployment{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      fixDeployName(),
-			Namespace: fixNamespaceName(),
-		},
-		Spec: api.DeploymentSpec{
-			Strategy: api.DeploymentStrategy{
-				RollingUpdate: &api.RollingUpdateDeployment{
-					MaxUnavailable: &intstr.IntOrString{IntVal: 1, StrVal: "1"},
-				},
-			},
-		},
-	}
-}
-
-func fixDeployWithPercentageMaxUnavailable() *api.Deployment {
-	return &api.Deployment{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      fixDeployName(),
-			Namespace: fixNamespaceName(),
-		},
-		Spec: api.DeploymentSpec{
-			Strategy: api.DeploymentStrategy{
-				RollingUpdate: &api.RollingUpdateDeployment{
-					MaxUnavailable: &intstr.IntOrString{StrVal: "20%"},
-				},
-			},
-		},
-	}
-}
-
-func fixDeployName() string {
-	return "deploy-name"
-}
-
-func ptrInt32(int int32) *int32 {
-	return &int
 }
